LibWeb: Move constructor and visit_edges in .cpp for EditingHostManager
